requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.20.0
openpyxl==3.0.6
pandas==1.2.1
pycodestyle==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
toml==0.10.2
xlrd==2.0.1
